+++ src/serialize_internal.h
@@ -304,6 +304,7 @@ void serialize_function(FILE *f, VirtualMachine *vm, SerializeContext *ctx,
 void serialize_function_signature(FILE *f, SerializeContext *ctx, Obj *fn);
 void serialize_global_var(FILE *f, VirtualMachine *vm, SerializeContext *ctx,
                           Obj *var);
+void serialize_dlfcn_shims(FILE *f, VirtualMachine *vm, Obj *prog);
 void serialize_native_accessor_shims(FILE *f, Obj *prog);
 void serialize_posix_compat_shims(FILE *f, VirtualMachine *vm, Obj *prog);
 void serialize_static_assert(FILE *f, VirtualMachine *vm, SerializeContext *ctx,